+++ docs/source/conf.py
@@ -12,13 +12,12 @@ import sys
 project = "Orichain"
 copyright = "2025, Oriserve"
 author = "Apoorv Singh, Shubham Maindola"
-release = "2.2.2"
+release = "2.2.3"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 extensions = []
-
 templates_path = ["_templates"]
 exclude_patterns = []
 
@@ -26,6 +25,7 @@ exclude_patterns = []
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
+html_show_sourcelink = False
 html_theme = "furo"
 html_static_path = ["_static"]
 html_css_files = ["custom.css"]
@@ -36,6 +36,7 @@ html_theme_options = {
 
 
 # -- Enabling sphinx.ext.autodoc (and others) to pull docstrings automatically -------------------------------------------------
+
 extensions = ["sphinx.ext.autodoc", "sphinx.ext.viewcode"]
 sys.path.insert(0, os.path.abspath("../../src"))
 autodoc_member_order = "bysource"