+++ src/orichain/llm/anthropic_llm.py
@@ -22,6 +22,7 @@ class Generate(object):
             - api_key (str): Anthropic API key
             - timeout (Timeout, optional): Request timeout parameter like connect, read, write. Default is 60.0, 5.0, 10.0, 2.0
             - max_retries (int, optional): Number of retries for the request. Default is 2
+            - prompt_caching (bool, optional): Whether to use prompt caching. Default: True
 
         Raises:
             - KeyError: If required parameters are not provided.
@@ -268,6 +269,7 @@ class AsyncGenerate(object):
             - api_key (str): Anthropic API key
             - timeout (Timeout, optional): Request timeout parameter like connect, read, write. Default is 60.0, 5.0, 10.0, 2.0
             - max_retries (int, optional): Number of retries for the request. Default is 2
+            - prompt_caching (bool, optional): Whether to use prompt caching. Default: True
 
         Raises:
             - KeyError: If required parameters are not provided.