+++ src/orichain/llm/openai_llm.py
@@ -79,7 +79,6 @@ class Generate(object):
             - sampling_paras (Optional[Dict], optional): Parameters for controlling the model's generation
             - system_prompt (Optional[str], optional): System prompt to provide context to the model
             - do_json (bool, optional): Whether to format the response as JSON. Defaults to False
-            - **kwds: Additional keyword arguments to pass to the client
 
         Returns:
             Dict: Response from the model or error information
@@ -327,7 +326,6 @@ class AsyncGenerate(object):
             - sampling_paras (Optional[Dict], optional): Parameters for controlling the model's generation
             - system_prompt (Optional[str], optional): System prompt to provide context to the model
             - do_json (bool, optional): Whether to format the response as JSON. Defaults to False
-            - **kwds: Additional keyword arguments to pass to the client
 
         Returns:
             Dict: Response from the model or error information