+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.3] - 2025-07-17
+
+### Changed
+- Updated the documentation.
+
+### Added
+- Added normalize_embeddings argument in sentence-transformers embedding models' call. 
+
+### Fixed
+- Removed request_id and uuid from Gemini & VertexAI chat formatters.
+
 ## [2.2.2] - 2025-07-15
 
 ### Changed